Implemented namespaces
A translation in the form of "User|Name" will use the full string as a key but only display the part after the pipe, so "Name". This is very helpful if you for example want to have "User|Name" translated differently to "Account|Name" in some languages but not all.